pyproject.toml
[project]
name = "passenv"
version = "0.2.1"
description = "Run commands with environment variables populated from pass entries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyyaml"]

[project.scripts]
passenv = "main:main"
